backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
python-multipart==0.0.9
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0

